add nav, nav-bottom, icons
