+++ App.js
@@ -1,23 +1,20 @@
 import React from "react";
+import { NavigationContainer } from "@react-navigation/native";
 import { useCallback } from "react";
-import {
-  Dimensions,
-  ImageBackground,
-  Platform,
-  StatusBar,
-  SafeAreaView,
-} from "react-native";
+import { Platform, SafeAreaView, StatusBar } from "react-native";
 import { useFonts } from "expo-font";
 import * as SplashScreen from "expo-splash-screen";
-import { styles } from "./styles";
 
-import LoginScreen from "./Screens/LoginScreen/LoginScreen";
-import RegistrationScreen from "./Screens/RegistrationScreen/RegistrationScreen";
+import { useRoute } from "./router";
+import { styles } from "./styles";
 
 SplashScreen.preventAutoHideAsync();
 
 export default function App() {
   console.log(Platform.OS);
+
+  const routing = useRoute(false);
+
   const [fontsLoaded] = useFonts({
     "Roboto-Regular": require("./assets/fonts/Roboto-Regular.ttf"),
     "Roboto-Medium": require("./assets/fonts/Roboto-Medium.ttf"),
@@ -34,20 +31,10 @@ export default function App() {
     return null;
   }
   return (
-    <>
-      <SafeAreaView style={styles.container} onLayout={onLayoutRootView}>
-        <ImageBackground
-          style={{
-            ...styles.image,
-            width: Dimensions.get("window").width,
-            height: Dimensions.get("window").height,
-          }}
-          source={require("./assets/bg.jpg")}
-        ></ImageBackground>
-        {true ? <RegistrationScreen /> : <LoginScreen />}
-        <StatusBar style="auto" />
-        {/* <StatusBar barStyle={"dark-content"} /> */}
-      </SafeAreaView>
-    </>
+    <SafeAreaView style={styles.container} onLayout={onLayoutRootView}>
+      <NavigationContainer>{routing}</NavigationContainer>
+      {/* {Platform.OS == "ios" ? <StatusBar barStyle={"dark-content"} :  <StatusBar style="auto" />} */}
+      <StatusBar style="auto" />
+    </SafeAreaView>
   );
 }
