+++ router.js
@@ -1,0 +1,74 @@
+import React from "react";
+
+import { createStackNavigator } from "@react-navigation/stack";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+// icons import
+import { AntDesign } from "@expo/vector-icons";
+import { Ionicons } from "@expo/vector-icons";
+
+import LoginScreen from "./screens/auth/LoginScreen";
+import RegistrationScreen from "./screens/auth/RegistrationScreen";
+import PostsScreen from "./screens/mainScreen/PostsScreen/PostsScreen";
+import CreatePostsScreen from "./screens/mainScreen/CreatePostsScreen/CreatePostsScreen";
+import ProfileScreen from "./screens/mainScreen/ProfileScreen/ProfileScreen";
+import Home from "./screens/Home/Home";
+
+const AuthStack = createStackNavigator();
+const MainTab = createBottomTabNavigator();
+
+export const useRoute = (isAuth) => {
+  if (!isAuth) {
+    return (
+      <AuthStack.Navigator>
+        <AuthStack.Screen
+          name="RegistrationScreen"
+          component={RegistrationScreen}
+          options={{ headerShown: false }}
+        />
+        <AuthStack.Screen
+          name="LoginScreen"
+          component={LoginScreen}
+          options={{ headerShown: false }}
+        />
+        <AuthStack.Screen
+          name="Home"
+          component={Home}
+          options={{ headerShown: false }}
+        />
+      </AuthStack.Navigator>
+    );
+  }
+  return (
+    <MainTab.Navigator tabBarOptions={{ showLabel: false }}>
+      <MainTab.Screen
+        options={{
+          tabBarIcon: ({ focused, size, color }) => (
+            <AntDesign name="appstore-o" size={size} color={color} />
+          ),
+        }}
+        name="Posts"
+        component={PostsScreen}
+      />
+
+      <MainTab.Screen
+        options={{
+          tabBarIcon: ({ focused, size, color }) => (
+            <Ionicons name="add-circle-sharp" size={size} color={color} />
+          ),
+        }}
+        name="Create"
+        component={CreatePostsScreen}
+      />
+
+      <MainTab.Screen
+        options={{
+          tabBarIcon: ({ focused, size, color }) => (
+            <Ionicons name="md-person-outline" size={size} color={color} />
+          ),
+        }}
+        name="Profile"
+        component={ProfileScreen}
+      />
+    </MainTab.Navigator>
+  );
+};
