+++ screens/Home/Home.js
@@ -1,0 +1,13 @@
+import React from "react";
+import { View, Text } from "react-native";
+import { styles } from "./styles";
+
+const Home = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Home Screen</Text>
+    </View>
+  );
+};
+
+export default Home;
