+++ screens/mainScreen/CreatePostsScreen/CreatePostsScreen.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+import { styles } from "./styles";
+
+const CreatePostsScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text>CreatePostsScreen</Text>
+    </View>
+  );
+};
+
+export default CreatePostsScreen;
