+++ screens/mainScreen/CreatePostsScreen/styles.js
@@ -3,8 +3,7 @@ import { StyleSheet } from "react-native";
 export const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: "#fff",
-    fontFamily: "Roboto-Regular",
-    justifyContent: "flex-end",
+    justifyContent: "center",
+    alignItems: "center",
   },
 });
