+++ screens/mainScreen/PostsScreen/PostsScreen.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+import { styles } from "./styles";
+
+const PostsScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text>PostsScreen</Text>
+    </View>
+  );
+};
+
+export default PostsScreen;
