+++ screens/mainScreen/ProfileScreen/ProfileScreen.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+import { styles } from "./styles";
+
+const ProfileScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text>ProfileScreen</Text>
+    </View>
+  );
+};
+
+export default ProfileScreen;
